+++ src/svp.cpp
@@ -1,6 +1,6 @@
 #include <math.h>
 
-double svp(double temp)
+float svp(float temp)
   /**********************************************************************
    This routine computes the saturated vapor pressure using Handbook
    of Hydrology eqn 4.2.2
@@ -9,11 +9,11 @@ double svp(double temp)
    
    **********************************************************************/
 {
-  double SVP;
+  float SVP;
   
-  double A_SVP=0.61078;
-  double B_SVP =17.269;
-  double C_SVP =237.3;
+  float A_SVP=0.61078;
+  float B_SVP =17.269;
+  float C_SVP =237.3;
   
   SVP = A_SVP * exp((B_SVP * temp)/(C_SVP+temp));
